Some refactoring, logic and prompt fixes
