+++ agents/test_case_classification/prompt.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from pathlib import Path
-from agents.prompt_base import PromptBase
+from common.prompt_base import PromptBase
 from common import utils
 
 logger = utils.get_logger("test_case_classification_prompt")
