+++ agents/test_case_generation/prompt_template.txt
@@ -37,4 +37,4 @@ Your tasks are the following, step-by-step:
     5.3. Double-check that all generated by you test steps adhere to the rules from "Test step generation rules" section. If not, correct non-adherent test steps so that they fully adhere to all those rules.
     5.4. Collect all generated by you test steps and generate only one test case which contains all those test steps, according to the rules from "Test case generation rules" section.
  6. Create generated by you test cases in the corresponding test management system using corresponding tool.
- 7. Return the generated by you test cases with your comments and updated test case IDs (keys) based on the result of the tool execution, as a final result in the requested format - don't execute any other tasks.
+ 7. Return the generated by you test cases having updated test case IDs (keys) based on the result of the tool execution, as a final result in the requested format - don't execute any other tasks.
