+++ agents/test_case_review/prompt_template.txt
@@ -9,7 +9,22 @@ Your tasks are the following:
       - Review the provided to you test case (taking into account all information which is present in the fetched by you media attachments) for coherence, redundancy, and effectiveness.
       - Based on your review, create a feedback which has only the most important explicit improvement suggestions in order to enhance the user story, thus addressing all your review comments.
       - Set the status of the test case to "Review Complete".
-   2. Wait for the status of all test cases to be updated, and after that add created by you feedbacks to the corresponding test cases using corresponding tool.
+   2. Wait for the status of all test cases to be updated, and after that add created by you feedbacks, formatted according to the 'Feedback Formatting Syntax' section, to the corresponding test cases using corresponding tool.
    3. After you've done processing all test cases and all corresponding tools have been executed, return created by you feedbacks as a final result, don't execute any other tasks.
 
 If you can't find any of the tools which are required in order to execute your tasks or if the tool returns execution results which are not expected by you - return immediately an error and interrupt execution.
+
+Feedback Formatting Syntax :
+     Bold text: **text** or __text__
+     Italics text: *text* or _text_
+     Strikethrough text: ~~text~~
+     Link: [link](www.atlassian.com)
+     Heading 1: #
+     Heading 2: ##
+     Heading 3: ###
+     Heading 4: ####
+     Unordered list: * or -
+     Ordered list: 1. or 2)
+     Block quote: >
+     Horizontal rule: --- or ***
+     Code block: ```
