+++ cloudbuild.yaml
@@ -3,6 +3,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 steps:
+  # Build the base image
+  - name: 'gcr.io/cloud-builders/docker'
+    args: [ 'build', '-t', 'agentic-qa-base:latest', '-f', 'Dockerfile.base', '.' ]
+
   # Build the orchestrator image
   - name: 'gcr.io/cloud-builders/docker'
     args: [ 'build', '-t', 'gcr.io/$PROJECT_ID/orchestrator', '-f', 'orchestrator/Dockerfile', '.' ]
@@ -51,6 +55,7 @@ steps:
       - '--platform=managed'
       - '--region=us-central1'
       - '--allow-unauthenticated'
+      - '--memory=1Gi'
       - '--max-instances=1'
       - '--set-secrets=JIRA_API_TOKEN=JIRA_API_TOKEN:latest,JIRA_USERNAME=JIRA_USERNAME:latest,JIRA_URL=JIRA_URL:latest'
       - '--port=9000'
@@ -73,7 +78,7 @@ steps:
       - '--vpc-egress=all-traffic'
       - '--region=us-central1'
       - '--allow-unauthenticated'
-      - '--memory=1Gi'
+      - '--memory=4Gi'
       - '--max-instances=1'
       - '--concurrency=1'
       - '--set-env-vars=^;^REMOTE_EXECUTION_AGENT_HOSTS=$_REMOTE_EXECUTION_AGENT_HOSTS'
@@ -96,6 +101,7 @@ steps:
       - '--vpc-egress=all-traffic'
       - '--region=us-central1'
       - '--allow-unauthenticated'
+      - '--memory=3Gi'
       - '--max-instances=1'
       - '--concurrency=1'
       - '--set-secrets=GOOGLE_API_KEY=GOOGLE_API_KEY:latest,JIRA_API_TOKEN=JIRA_API_TOKEN:latest,JIRA_USERNAME=JIRA_USERNAME:latest,ZEPHYR_API_TOKEN=ZEPHYR_API_TOKEN:latest,JIRA_URL=JIRA_URL:latest,JIRA_MCP_SERVER_URL=JIRA_MCP_SERVER_URL:latest,ZEPHYR_BASE_URL=ZEPHYR_BASE_URL:latest'
@@ -115,6 +121,7 @@ steps:
       - '--vpc-egress=all-traffic'
       - '--region=us-central1'
       - '--allow-unauthenticated'
+      - '--memory=3Gi'
       - '--max-instances=1'
       - '--concurrency=1'
       - '--set-secrets=GOOGLE_API_KEY=GOOGLE_API_KEY:latest,JIRA_API_TOKEN=JIRA_API_TOKEN:latest,JIRA_USERNAME=JIRA_USERNAME:latest,ZEPHYR_API_TOKEN=ZEPHYR_API_TOKEN:latest,JIRA_URL=JIRA_URL:latest,JIRA_MCP_SERVER_URL=JIRA_MCP_SERVER_URL:latest,ZEPHYR_BASE_URL=ZEPHYR_BASE_URL:latest'
@@ -134,6 +141,7 @@ steps:
       - '--vpc-egress=all-traffic'
       - '--region=us-central1'
       - '--allow-unauthenticated'
+      - '--memory=3Gi'
       - '--max-instances=1'
       - '--concurrency=1'
       - '--set-secrets=GOOGLE_API_KEY=GOOGLE_API_KEY:latest,JIRA_API_TOKEN=JIRA_API_TOKEN:latest,JIRA_USERNAME=JIRA_USERNAME:latest,ZEPHYR_API_TOKEN=ZEPHYR_API_TOKEN:latest,JIRA_URL=JIRA_URL:latest,JIRA_MCP_SERVER_URL=JIRA_MCP_SERVER_URL:latest,ZEPHYR_BASE_URL=ZEPHYR_BASE_URL:latest'
@@ -153,6 +161,7 @@ steps:
       - '--vpc-egress=all-traffic'
       - '--region=us-central1'
       - '--allow-unauthenticated'
+      - '--memory=3Gi'
       - '--max-instances=1'
       - '--concurrency=1'
       - '--set-secrets=GOOGLE_API_KEY=GOOGLE_API_KEY:latest,JIRA_API_TOKEN=JIRA_API_TOKEN:latest,JIRA_USERNAME=JIRA_USERNAME:latest,ZEPHYR_API_TOKEN=ZEPHYR_API_TOKEN:latest,JIRA_URL=JIRA_URL:latest,JIRA_MCP_SERVER_URL=JIRA_MCP_SERVER_URL:latest,ZEPHYR_BASE_URL=ZEPHYR_BASE_URL:latest'
