+++ orchestrator/Dockerfile
@@ -1,14 +1,14 @@
 # SPDX-FileCopyrightText: 2025 Taras Paruta (partarstu@gmail.com)
 #
 # SPDX-License-Identifier: Apache-2.0
-FROM python:3.13-slim
+FROM agentic-qa-base:latest
 
 ENV DEBIAN_FRONTEND=noninteractive
 
 # Create man directories to prevent potential dependency issues with ca-certificates
 RUN mkdir -p /usr/share/man/man1 /usr/share/man/man2 && \
     apt-get update && \
-    apt-get install -y --no-install-recommends openjdk-17-jre-headless wget unzip && \
+    apt-get install -y --no-install-recommends openjdk-25-jre-headless wget unzip && \
     apt-get clean && rm -rf /var/lib/apt/lists/*
 
 # Set the Allure version as an argument
@@ -27,12 +27,11 @@ RUN wget -q https://github.com/allure-framework/allure2/releases/download/${ALLU
 RUN allure --version
 RUN mkdir "allure-report"
 
-WORKDIR /app
+ARG WORK_DIR=/app
+WORKDIR ${WORK_DIR}
 
-COPY requirements.txt .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
+COPY orchestrator/ ${WORK_DIR}/orchestrator
+COPY common/ ${WORK_DIR}/common
+COPY config.py ${WORK_DIR}/config.py
 
 CMD ["gunicorn", "-w", "1", "-k", "uvicorn.workers.UvicornWorker", "orchestrator.main:orchestrator_app"]
